Change mutation logic.
Now mutations are not perforemd for each triangle with a different
probability for each mutation, but there is a per-triangle mutation
probability, and when a triangle is mutated, just a single mutation is
picked. The mutation probability of a trinagle can now be adjusted via
the command line parameter --mutation-rate.

This is an improvements because:

1. Images obtained this way find the same level of solution a bit faster
apparently.
2. They have more "clean" look, even when similarly fit.
3. With the new logic there is a single parameter to expose to the user.